+++ apps/src/codebridge/InfoPanel/InfoPanel.tsx
