+++ apps/src/codebridge/Workspace/Workspace.tsx
@@ -1,12 +1,12 @@
 import Alert from '@code-dot-org/component-library/alert';
 import {useCodebridgeContext} from '@codebridge/codebridgeContext';
 import ToggleFileBrowserButton from '@codebridge/components/ToggleFileBrowserButton';
-import {FileTabs} from '@codebridge/FileTabs';
 import classnames from 'classnames';
 import React from 'react';
 
 import {Editor} from '@cdo/apps/codebridge/Editor/Editor';
 import {FileBrowser} from '@cdo/apps/codebridge/FileBrowser/FileBrowser';
+import {FileTabs} from '@cdo/apps/codebridge/FileTabs/FileTabs';
 import codebridgeI18n from '@cdo/apps/codebridge/locale';
 import {START_SOURCES, WARNING_BANNER_MESSAGES} from '@cdo/apps/lab2/constants';
 import {isProjectTemplateLevel} from '@cdo/apps/lab2/lab2Redux';
